backend/requirements.txt
django
google-generativeai
python-dotenv
django-cors-headers
gitpython
djangorestframework
